Create SimonSaysRev2
This version was an attempt to implement while loops as a way of ending some issues with user inputs or the lack there of. It was also an attempt to implement an endGame sequence. The major problems with this version was some issues with delays making the game unplayable and an issue where the user would have to input the correct sequence as many times as how far you were in the sequence. For example, if you were on the fourth number in the sequence, in order to get to the next sequence you would have to input the correct pattern 4 times. When you made it to 5, you would have to input the sequence 5 times before going to 6 and so on. This issue was solved by getting rid of the while loop found at line 119.